requirements.txt
langchain==0.2.5
langchain-community==0.2.5
langchain-anthropic==0.1.15
fastapi==0.111.0
uvicorn==0.30.1
pyyaml==6.0.1
torch==2.2.0
transformers==4.41.2